+++ mayan/apps/testing/tests/base.py
@@ -7,6 +7,10 @@ from django_test_migrations.contrib.unittest_case import MigratorTestCase
 
 from mayan.apps.acls.tests.mixins import ACLTestCaseMixin
 from mayan.apps.converter.tests.mixins import LayerTestCaseMixin
+from mayan.apps.databases.tests.mixins import (
+    ConnectionsCheckTestCaseMixin, ModelTestCaseMixin,
+    TestModelTestCaseMixin, RandomPrimaryKeyModelMonkeyPatchMixin
+)
 from mayan.apps.events.tests.mixins.event_mixins import EventTestCaseMixin
 from mayan.apps.logging.tests.mixins import TestCaseMixinSilenceLogger
 from mayan.apps.permissions.tests.mixins import PermissionTestCaseMixin
@@ -19,10 +23,9 @@ from mayan.apps.user_management.tests.mixins.user_mixins import UserTestMixin
 from ..literals import EXCLUDE_TEST_TAG
 
 from .mixins import (
-    ConnectionsCheckTestCaseMixin, ContentTypeCheckTestCaseMixin,
-    DelayTestCaseMixin, DescriptorLeakCheckTestCaseMixin, ModelTestCaseMixin,
-    OpenFileCheckTestCaseMixin, RandomPrimaryKeyModelMonkeyPatchMixin,
-    TempfileCheckTestCasekMixin, TestModelTestCaseMixin
+    ContentTypeCheckTestCaseMixin, DelayTestCaseMixin,
+    DescriptorLeakCheckTestCaseMixin, OpenFileCheckTestCaseMixin,
+    TempfileCheckTestCasekMixin
 )